+++ test-output/Default suite/Default test.html
@@ -55,11 +55,11 @@ function toggleAllBoxes() {
 <body>
 <h2 align='center'>Default test</h2><table border='1' align="center">
 <tr>
-<td>Tests passed/Failed/Skipped:</td><td>0/1/0</td>
+<td>Tests passed/Failed/Skipped:</td><td>1/0/0</td>
 </tr><tr>
-<td>Started on:</td><td>Mon Apr 28 11:44:16 IST 2025</td>
+<td>Started on:</td><td>Tue Apr 29 12:08:13 IST 2025</td>
 </tr>
-<tr><td>Total time:</td><td>51 seconds (51106 ms)</td>
+<tr><td>Total time:</td><td>116 seconds (116076 ms)</td>
 </tr><tr>
 <td>Included groups:</td><td></td>
 </tr><tr>
@@ -67,8 +67,8 @@ function toggleAllBoxes() {
 </tr>
 </table><p/>
 <small><i>(Hover the method name to see the test class name)</i></small><p/>
-<table width='100%' border='1' class='invocation-failed'>
-<tr><td colspan='4' align='center'><b>FAILED TESTS</b></td></tr>
+<table width='100%' border='1' class='invocation-passed'>
+<tr><td colspan='4' align='center'><b>PASSED TESTS</b></td></tr>
 <tr><td><b>Test method</b></td>
 <td><b>Attribute(s)</b></td>
 <td width="30%"><b>Exception</b></td>
@@ -76,81 +76,10 @@ function toggleAllBoxes() {
 <td><b>Instance</b></td>
 </tr>
 <tr>
-<td title='petsitcare_Admin_TestCases.ManageSubscriberTestCase.testSubscribersActions()'><b>testSubscribersActions</b><br>Test class: petsitcare_Admin_TestCases.ManageSubscriberTestCase</td>
-<td><div><pre>java.io.IOException: Cannot create directory &apos;C:\Users\NISHITH\git\Petsitcare35\screenshots&apos;.
-	at org.apache.commons.io.FileUtils.mkdirs(FileUtils.java:2289)
-	at org.apache.commons.io.FileUtils.createParentDirectories(FileUtils.java:1097)
-	at org.apache.commons.io.FileUtils.copyFile(FileUtils.java:802)
-	at org.apache.commons.io.FileUtils.copyFile(FileUtils.java:834)
-	at org.apache.commons.io.FileUtils.copyFile(FileUtils.java:742)
-	at petsitcare_Admin_base.Base.screenshot(Base.java:47)
-	at petsitcare_Admin_TestCases.login.login1(login.java:43)
-	at petsitcare_Admin_TestCases.ManageSubscriberTestCase.testSubscribersActions(ManageSubscriberTestCase.java:22)
-	at java.base/jdk.internal.reflect.DirectMethodHandleAccessor.invoke(DirectMethodHandleAccessor.java:103)
-	at java.base/java.lang.reflect.Method.invoke(Method.java:580)
-	at org.testng.internal.invokers.MethodInvocationHelper.invokeMethod(MethodInvocationHelper.java:141)
-	at org.testng.internal.invokers.TestInvoker.invokeMethod(TestInvoker.java:686)
-	at org.testng.internal.invokers.TestInvoker.invokeTestMethod(TestInvoker.java:230)
-	at org.testng.internal.invokers.MethodRunner.runInSequence(MethodRunner.java:63)
-	at org.testng.internal.invokers.TestInvoker$MethodInvocationAgent.invoke(TestInvoker.java:992)
-	at org.testng.internal.invokers.TestInvoker.invokeTestMethods(TestInvoker.java:203)
-	at org.testng.internal.invokers.TestMethodWorker.invokeTestMethods(TestMethodWorker.java:154)
-	at org.testng.internal.invokers.TestMethodWorker.run(TestMethodWorker.java:134)
-	at java.base/java.util.ArrayList.forEach(ArrayList.java:1597)
-	at org.testng.TestRunner.privateRun(TestRunner.java:739)
-	at org.testng.TestRunner.run(TestRunner.java:614)
-	at org.testng.SuiteRunner.runTest(SuiteRunner.java:421)
-	at org.testng.SuiteRunner.runSequentially(SuiteRunner.java:413)
-	at org.testng.SuiteRunner.privateRun(SuiteRunner.java:373)
-	at org.testng.SuiteRunner.run(SuiteRunner.java:312)
-	at org.testng.SuiteRunnerWorker.runSuite(SuiteRunnerWorker.java:52)
-	at org.testng.SuiteRunnerWorker.run(SuiteRunnerWorker.java:95)
-	at org.testng.TestNG.runSuitesSequentially(TestNG.java:1274)
-	at org.testng.TestNG.runSuitesLocally(TestNG.java:1208)
-	at org.testng.TestNG.runSuites(TestNG.java:1112)
-	at org.testng.TestNG.run(TestNG.java:1079)
-	at org.testng.remote.AbstractRemoteTestNG.run(AbstractRemoteTestNG.java:115)
-	at org.testng.remote.RemoteTestNG.initAndRun(RemoteTestNG.java:293)
-	at org.testng.remote.RemoteTestNG.main(RemoteTestNG.java:91)
-</pre></div><a href='#' onClick='toggleBox("stack-trace208301085", this, "Click to show all stack frames", "Click to hide stack frames")'>Click to show all stack frames</a>
-<div class='stack-trace' id='stack-trace208301085'><pre>java.io.IOException: Cannot create directory &apos;C:\Users\NISHITH\git\Petsitcare35\screenshots&apos;.
-	at org.apache.commons.io.FileUtils.mkdirs(FileUtils.java:2289)
-	at org.apache.commons.io.FileUtils.createParentDirectories(FileUtils.java:1097)
-	at org.apache.commons.io.FileUtils.copyFile(FileUtils.java:802)
-	at org.apache.commons.io.FileUtils.copyFile(FileUtils.java:834)
-	at org.apache.commons.io.FileUtils.copyFile(FileUtils.java:742)
-	at petsitcare_Admin_base.Base.screenshot(Base.java:47)
-	at petsitcare_Admin_TestCases.login.login1(login.java:43)
-	at petsitcare_Admin_TestCases.ManageSubscriberTestCase.testSubscribersActions(ManageSubscriberTestCase.java:22)
-	at java.base/jdk.internal.reflect.DirectMethodHandleAccessor.invoke(DirectMethodHandleAccessor.java:103)
-	at java.base/java.lang.reflect.Method.invoke(Method.java:580)
-	at org.testng.internal.invokers.MethodInvocationHelper.invokeMethod(MethodInvocationHelper.java:141)
-	at org.testng.internal.invokers.TestInvoker.invokeMethod(TestInvoker.java:686)
-	at org.testng.internal.invokers.TestInvoker.invokeTestMethod(TestInvoker.java:230)
-	at org.testng.internal.invokers.MethodRunner.runInSequence(MethodRunner.java:63)
-	at org.testng.internal.invokers.TestInvoker$MethodInvocationAgent.invoke(TestInvoker.java:992)
-	at org.testng.internal.invokers.TestInvoker.invokeTestMethods(TestInvoker.java:203)
-	at org.testng.internal.invokers.TestMethodWorker.invokeTestMethods(TestMethodWorker.java:154)
-	at org.testng.internal.invokers.TestMethodWorker.run(TestMethodWorker.java:134)
-	at java.base/java.util.ArrayList.forEach(ArrayList.java:1597)
-	at org.testng.TestRunner.privateRun(TestRunner.java:739)
-	at org.testng.TestRunner.run(TestRunner.java:614)
-	at org.testng.SuiteRunner.runTest(SuiteRunner.java:421)
-	at org.testng.SuiteRunner.runSequentially(SuiteRunner.java:413)
-	at org.testng.SuiteRunner.privateRun(SuiteRunner.java:373)
-	at org.testng.SuiteRunner.run(SuiteRunner.java:312)
-	at org.testng.SuiteRunnerWorker.runSuite(SuiteRunnerWorker.java:52)
-	at org.testng.SuiteRunnerWorker.run(SuiteRunnerWorker.java:95)
-	at org.testng.TestNG.runSuitesSequentially(TestNG.java:1274)
-	at org.testng.TestNG.runSuitesLocally(TestNG.java:1208)
-	at org.testng.TestNG.runSuites(TestNG.java:1112)
-	at org.testng.TestNG.run(TestNG.java:1079)
-	at org.testng.remote.AbstractRemoteTestNG.run(AbstractRemoteTestNG.java:115)
-	at org.testng.remote.RemoteTestNG.initAndRun(RemoteTestNG.java:293)
-	at org.testng.remote.RemoteTestNG.main(RemoteTestNG.java:91)
-</pre></div></td>
-<td>24</td>
-<td>petsitcare_Admin_TestCases.ManageSubscriberTestCase@37ceb1df</td></tr>
+<td title='petsitcare_Admin_TestCases.ManageSubscribersTest.testSubscribersActions()'><b>testSubscribersActions</b><br>Test class: petsitcare_Admin_TestCases.ManageSubscribersTest</td>
+<td></td>
+<td>93</td>
+<td>petsitcare_Admin_TestCases.ManageSubscribersTest@37ceb1df</td></tr>
 </table><p>
 </body>
 </html>
