+++ test-output/Default suite/Default test.xml
@@ -1,44 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitXMLReporter -->
-<testsuite ignored="0" hostname="DESKTOP-ERTMBQN" failures="1" tests="1" name="Default test" time="51.106" errors="0" timestamp="2025-04-28T11:45:08 IST">
-  <testcase classname="petsitcare_Admin_TestCases.ManageSubscriberTestCase" name="testSubscribersActions" time="24.326">
-    <failure type="java.io.IOException" message="Cannot create directory &amp;apos;C:\Users\NISHITH\git\Petsitcare35\screenshots&amp;apos;.">
-      <![CDATA[java.io.IOException: Cannot create directory 'C:\Users\NISHITH\git\Petsitcare35\screenshots'.
-at org.apache.commons.io.FileUtils.mkdirs(FileUtils.java:2289)
-at org.apache.commons.io.FileUtils.createParentDirectories(FileUtils.java:1097)
-at org.apache.commons.io.FileUtils.copyFile(FileUtils.java:802)
-at org.apache.commons.io.FileUtils.copyFile(FileUtils.java:834)
-at org.apache.commons.io.FileUtils.copyFile(FileUtils.java:742)
-at petsitcare_Admin_base.Base.screenshot(Base.java:47)
-at petsitcare_Admin_TestCases.login.login1(login.java:43)
-at petsitcare_Admin_TestCases.ManageSubscriberTestCase.testSubscribersActions(ManageSubscriberTestCase.java:22)
-at java.base/jdk.internal.reflect.DirectMethodHandleAccessor.invoke(DirectMethodHandleAccessor.java:103)
-at java.base/java.lang.reflect.Method.invoke(Method.java:580)
-at org.testng.internal.invokers.MethodInvocationHelper.invokeMethod(MethodInvocationHelper.java:141)
-at org.testng.internal.invokers.TestInvoker.invokeMethod(TestInvoker.java:686)
-at org.testng.internal.invokers.TestInvoker.invokeTestMethod(TestInvoker.java:230)
-at org.testng.internal.invokers.MethodRunner.runInSequence(MethodRunner.java:63)
-at org.testng.internal.invokers.TestInvoker$MethodInvocationAgent.invoke(TestInvoker.java:992)
-at org.testng.internal.invokers.TestInvoker.invokeTestMethods(TestInvoker.java:203)
-at org.testng.internal.invokers.TestMethodWorker.invokeTestMethods(TestMethodWorker.java:154)
-at org.testng.internal.invokers.TestMethodWorker.run(TestMethodWorker.java:134)
-at java.base/java.util.ArrayList.forEach(ArrayList.java:1597)
-at org.testng.TestRunner.privateRun(TestRunner.java:739)
-at org.testng.TestRunner.run(TestRunner.java:614)
-at org.testng.SuiteRunner.runTest(SuiteRunner.java:421)
-at org.testng.SuiteRunner.runSequentially(SuiteRunner.java:413)
-at org.testng.SuiteRunner.privateRun(SuiteRunner.java:373)
-at org.testng.SuiteRunner.run(SuiteRunner.java:312)
-at org.testng.SuiteRunnerWorker.runSuite(SuiteRunnerWorker.java:52)
-at org.testng.SuiteRunnerWorker.run(SuiteRunnerWorker.java:95)
-at org.testng.TestNG.runSuitesSequentially(TestNG.java:1274)
-at org.testng.TestNG.runSuitesLocally(TestNG.java:1208)
-at org.testng.TestNG.runSuites(TestNG.java:1112)
-at org.testng.TestNG.run(TestNG.java:1079)
-at org.testng.remote.AbstractRemoteTestNG.run(AbstractRemoteTestNG.java:115)
-at org.testng.remote.RemoteTestNG.initAndRun(RemoteTestNG.java:293)
-at org.testng.remote.RemoteTestNG.main(RemoteTestNG.java:91)
-]]>
-    </failure>
-  </testcase> <!-- testSubscribersActions -->
+<testsuite ignored="0" hostname="DESKTOP-ERTMBQN" failures="0" tests="1" name="Default test" time="116.076" errors="0" timestamp="2025-04-29T12:10:09 IST">
+  <testcase classname="petsitcare_Admin_TestCases.ManageSubscribersTest" name="testSubscribersActions" time="93.453"/>
 </testsuite> <!-- Default test -->
