+++ test-output/Suite/Test.html
@@ -57,9 +57,9 @@ function toggleAllBoxes() {
 <tr>
 <td>Tests passed/Failed/Skipped:</td><td>1/0/0</td>
 </tr><tr>
-<td>Started on:</td><td>Mon Apr 28 12:01:04 IST 2025</td>
+<td>Started on:</td><td>Wed Apr 30 11:37:01 IST 2025</td>
 </tr>
-<tr><td>Total time:</td><td>116 seconds (116945 ms)</td>
+<tr><td>Total time:</td><td>109 seconds (109467 ms)</td>
 </tr><tr>
 <td>Included groups:</td><td></td>
 </tr><tr>
@@ -76,10 +76,10 @@ function toggleAllBoxes() {
 <td><b>Instance</b></td>
 </tr>
 <tr>
-<td title='petsitcare_Admin_TestCases.ManageSubscriberTestCase.testSubscribersActions()'><b>testSubscribersActions</b><br>Test class: petsitcare_Admin_TestCases.ManageSubscriberTestCase</td>
+<td title='petsitcare_Admin_TestCases.ManageSubscribersTest.testSubscribersActions()'><b>testSubscribersActions</b><br>Test class: petsitcare_Admin_TestCases.ManageSubscribersTest</td>
 <td></td>
-<td>98</td>
-<td>petsitcare_Admin_TestCases.ManageSubscriberTestCase@37ceb1df</td></tr>
+<td>91</td>
+<td>petsitcare_Admin_TestCases.ManageSubscribersTest@37ceb1df</td></tr>
 </table><p>
 </body>
 </html>
