+++ test-output/Suite/Test.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitXMLReporter -->
-<testsuite ignored="0" hostname="DESKTOP-ERTMBQN" failures="0" tests="1" name="Test" time="116.945" errors="0" timestamp="2025-04-28T12:03:01 IST">
-  <testcase classname="petsitcare_Admin_TestCases.ManageSubscriberTestCase" name="testSubscribersActions" time="98.433"/>
+<testsuite ignored="0" hostname="DESKTOP-ERTMBQN" failures="0" tests="1" name="Test" time="109.467" errors="0" timestamp="2025-04-30T11:38:50 IST">
+  <testcase classname="petsitcare_Admin_TestCases.ManageSubscribersTest" name="testSubscribersActions" time="91.957"/>
 </testsuite> <!-- Test -->
