+++ test-output/index.html
@@ -103,7 +103,7 @@
                     <div class="method-list-content passed suite-Suite">
                       <span>
                         <img src="passed.png" width="3%"/>
-                        <a href="#" panel-name="suite-Suite" title="petsitcare_Admin_TestCases.ManageSubscriberTestCase" class="method navigator-link" hash-for-method="testSubscribersActions">testSubscribersActions</a> <!-- method navigator-link -->
+                        <a href="#" panel-name="suite-Suite" title="petsitcare_Admin_TestCases.ManageSubscribersTest" class="method navigator-link" hash-for-method="testSubscribersActions">testSubscribersActions</a> <!-- method navigator-link -->
                       </span>
                       <br/>
                     </div> <!-- method-list-content passed suite-Suite -->
@@ -121,7 +121,7 @@
           <div class="suite-Suite-class-passed">
             <div class="main-panel-header rounded-window-top">
               <img src="passed.png"/>
-              <span class="class-name">petsitcare_Admin_TestCases.ManageSubscriberTestCase</span>
+              <span class="class-name">petsitcare_Admin_TestCases.ManageSubscribersTest</span>
             </div> <!-- main-panel-header rounded-window-top -->
             <div class="main-panel-content rounded-window-bottom">
               <div class="method">
@@ -145,7 +145,7 @@
 &lt;suite name=&quot;Suite&quot; guice-stage=&quot;DEVELOPMENT&quot;&gt;
   &lt;test thread-count=&quot;5&quot; name=&quot;Test&quot;&gt;
     &lt;classes&gt;
-      &lt;class name=&quot;petsitcare_Admin_TestCases.ManageSubscriberTestCase&quot;/&gt;
+      &lt;class name=&quot;petsitcare_Admin_TestCases.ManageSubscribersTest&quot;/&gt;
     &lt;/classes&gt;
   &lt;/test&gt; &lt;!-- Test --&gt;
 &lt;/suite&gt; &lt;!-- Suite --&gt;
@@ -188,8 +188,8 @@ data.addColumn('number', 'Time (ms)');
 data.addRows(1);
 data.setCell(0, 0, 0)
 data.setCell(0, 1, 'testSubscribersActions')
-data.setCell(0, 2, 'petsitcare_Admin_TestCases.ManageSubscriberTestCase')
-data.setCell(0, 3, 98433);
+data.setCell(0, 2, 'petsitcare_Admin_TestCases.ManageSubscribersTest')
+data.setCell(0, 3, 91957);
 window.suiteTableData['Suite']= { tableData: data, tableDiv: 'times-div-Suite'}
 return data;
 }
@@ -220,18 +220,18 @@ return data;
           </div> <!-- main-panel-header rounded-window-top -->
           <div class="main-panel-content rounded-window-bottom">
             <div class="chronological-class">
-              <div class="chronological-class-name">petsitcare_Admin_TestCases.ManageSubscriberTestCase</div> <!-- chronological-class-name -->
+              <div class="chronological-class-name">petsitcare_Admin_TestCases.ManageSubscribersTest</div> <!-- chronological-class-name -->
               <div class="configuration-class before">
                 <span class="method-name">openurl</span>
                 <span class="method-start">0 ms</span>
               </div> <!-- configuration-class before -->
               <div class="test-method">
                 <span class="method-name">testSubscribersActions</span>
-                <span class="method-start">17378 ms</span>
+                <span class="method-start">15854 ms</span>
               </div> <!-- test-method -->
               <div class="configuration-class after">
                 <span class="method-name">closeurl</span>
-                <span class="method-start">115845 ms</span>
+                <span class="method-start">107873 ms</span>
               </div> <!-- configuration-class after -->
           </div> <!-- main-panel-content rounded-window-bottom -->
         </div> <!-- panel -->
