+++ test-output/junitreports/TEST-petsitcare_Admin_TestCases.ManageSubscribersTest.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Generated by org.testng.reporters.JUnitReportReporter -->
+<testsuite hostname="DESKTOP-ERTMBQN" failures="0" tests="1" name="petsitcare_Admin_TestCases.ManageSubscribersTest" time="91.957" errors="0" timestamp="2025-04-30T11:38:51 IST" skipped="0">
+  <testcase classname="petsitcare_Admin_TestCases.ManageSubscribersTest" name="testSubscribersActions" time="91.957"/>
+</testsuite> <!-- petsitcare_Admin_TestCases.ManageSubscribersTest -->
