+++ test-output/junitreports/TEST-petsitcare_Admin_TestCases.login.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitReportReporter -->
-<testsuite hostname="DESKTOP-APLM1O2" failures="0" tests="1" name="petsitcare_Admin_TestCases.login" time="26.743" errors="0" timestamp="2025-04-10T10:31:56 IST" skipped="0">
-  <testcase classname="petsitcare_Admin_TestCases.login" name="login1" time="26.743"/>
+<testsuite hostname="DESKTOP-ERTMBQN" failures="0" tests="1" name="petsitcare_Admin_TestCases.login" time="63.941" errors="0" timestamp="2025-04-29T10:57:25 IST" skipped="0">
+  <testcase classname="petsitcare_Admin_TestCases.login" name="login1" time="63.941"/>
 </testsuite> <!-- petsitcare_Admin_TestCases.login -->
