+++ test-output/testng-results.xml
@@ -2,24 +2,24 @@
 <testng-results ignored="0" total="1" passed="1" failed="0" skipped="0">
   <reporter-output>
   </reporter-output>
-  <suite started-at="2025-04-28T12:01:04 IST" name="Suite" finished-at="2025-04-28T12:03:01 IST" duration-ms="116945">
+  <suite started-at="2025-04-30T11:37:01 IST" name="Suite" finished-at="2025-04-30T11:38:50 IST" duration-ms="109467">
     <groups>
     </groups>
-    <test started-at="2025-04-28T12:01:04 IST" name="Test" finished-at="2025-04-28T12:03:01 IST" duration-ms="116945">
-      <class name="petsitcare_Admin_TestCases.ManageSubscriberTestCase">
-        <test-method is-config="true" signature="openurl()[pri:0, instance:petsitcare_Admin_TestCases.ManageSubscriberTestCase@37ceb1df]" started-at="2025-04-28T12:01:04 IST" name="openurl" finished-at="2025-04-28T12:01:22 IST" duration-ms="17243" status="PASS">
+    <test started-at="2025-04-30T11:37:01 IST" name="Test" finished-at="2025-04-30T11:38:50 IST" duration-ms="109467">
+      <class name="petsitcare_Admin_TestCases.ManageSubscribersTest">
+        <test-method is-config="true" signature="openurl()[pri:0, instance:petsitcare_Admin_TestCases.ManageSubscribersTest@37ceb1df]" started-at="2025-04-30T11:37:01 IST" name="openurl" finished-at="2025-04-30T11:37:17 IST" duration-ms="15730" status="PASS">
           <reporter-output>
           </reporter-output>
         </test-method> <!-- openurl -->
-        <test-method signature="testSubscribersActions()[pri:0, instance:petsitcare_Admin_TestCases.ManageSubscriberTestCase@37ceb1df]" started-at="2025-04-28T12:01:22 IST" name="testSubscribersActions" finished-at="2025-04-28T12:03:00 IST" duration-ms="98433" status="PASS">
+        <test-method signature="testSubscribersActions()[pri:0, instance:petsitcare_Admin_TestCases.ManageSubscribersTest@37ceb1df]" started-at="2025-04-30T11:37:17 IST" name="testSubscribersActions" finished-at="2025-04-30T11:38:49 IST" duration-ms="91957" status="PASS">
           <reporter-output>
           </reporter-output>
         </test-method> <!-- testSubscribersActions -->
-        <test-method is-config="true" signature="closeurl()[pri:0, instance:petsitcare_Admin_TestCases.ManageSubscriberTestCase@37ceb1df]" started-at="2025-04-28T12:03:00 IST" name="closeurl" finished-at="2025-04-28T12:03:01 IST" duration-ms="963" status="PASS">
+        <test-method is-config="true" signature="closeurl()[pri:0, instance:petsitcare_Admin_TestCases.ManageSubscribersTest@37ceb1df]" started-at="2025-04-30T11:38:49 IST" name="closeurl" finished-at="2025-04-30T11:38:50 IST" duration-ms="1480" status="PASS">
           <reporter-output>
           </reporter-output>
         </test-method> <!-- closeurl -->
-      </class> <!-- petsitcare_Admin_TestCases.ManageSubscriberTestCase -->
+      </class> <!-- petsitcare_Admin_TestCases.ManageSubscribersTest -->
     </test> <!-- Test -->
   </suite> <!-- Suite -->
 </testng-results>
